LPS-8302 Backend
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@48059 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/editor/fckeditor/receiver/impl/BaseCommandReceiver.java
@@ -281,7 +281,7 @@ public abstract class BaseCommandReceiver implements CommandReceiver {
 		return getSize(0);
 	}
 
-	protected String getSize(int size) {
+	protected String getSize(long size) {
 		return String.valueOf(Math.ceil(size / 1000));
 	}
