+++ portal-impl/src/com/liferay/portal/sharepoint/BaseSharepointStorageImpl.java
@@ -132,7 +132,7 @@ public abstract class BaseSharepointStorageImpl implements SharepointStorage {
 	}
 
 	protected Tree getDocumentTree(
-		String documentName, Date createDate, Date modifiedDate, int size,
+		String documentName, Date createDate, Date modifiedDate, long size,
 		String userName, String version) {
 
 		Tree documentTree = new Tree();
