+++ portal-impl/src/com/liferay/portlet/documentlibrary/action/GetFileAction.java
@@ -243,14 +243,14 @@ public class GetFileAction extends PortletAction {
 
 		if (!converted) {
 			if (DLUtil.compareVersions(version, fileEntry.getVersion()) >= 0) {
-				contentLength = fileEntry.getSize();
+				contentLength = (int)fileEntry.getSize();
 			}
 			else {
 				DLFileVersion fileVersion =
 					DLFileVersionLocalServiceUtil.getFileVersion(
 						groupId, folderId, name, version);
 
-				contentLength = fileVersion.getSize();
+				contentLength = (int)fileVersion.getSize();
 			}
 		}
